Creates Intro component and moves intro section. Minor refactoring.
