+++ src/Components/Technologies.jsx
@@ -41,6 +41,13 @@ class Technologies extends Component {
 
     return (
       <>
+        <div id="technologies">
+          <p id="tech-paragraph">
+            These are some of the
+            <br />
+            technologies & methodologies I've worked with
+          </p>
+        </div>
         <Container id="tech-container">
           <div id="framework-list">
             <h1 className="list-header">Frameworks / Libraries</h1>
